+++ src/main/scala/examples/oauth/User.scala
@@ -1,3 +1,0 @@
-package examples.oauth
-
-case class User(name: String)